refactor(tttm): move the binary to its own tttm crate
