+++ src/tttm/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tttm"
+authors.workspace = true
+categories.workspace = true
+description.workspace = true
+documentation.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[[bin]]
+name = "tttm"
+path = "src/main.rs"
