+++ src/tttm/src/components/square.rs
