+++ src/tttm/src/components/square_value.rs
