+++ src/tttm/src/gameplay/errors.rs
